Simplify for interim release
