--- KoAR.SaveEditor/Themes/Generic.ReleaseDownloader.xaml
@@ -1,5 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:KoAR.SaveEditor.Views.Updates">
-    
-</ResourceDictionary>
