+++ KoAR.SaveEditor/Views/Updates/UpdateControl.xaml
@@ -1,0 +1,121 @@
+﻿<UserControl x:Class="KoAR.SaveEditor.Views.Updates.UpdateControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+    xmlns:editor="clr-namespace:KoAR.SaveEditor"
+    xmlns:local="clr-namespace:KoAR.SaveEditor.Views.Updates"
+    xmlns:constructs="clr-namespace:KoAR.SaveEditor.Constructs"
+    mc:Ignorable="d" d:DataContext="{d:DesignInstance {x:Type local:UpdateViewModelBase}}">
+    <UserControl.Resources>
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="/KoAR.SaveEditor;component/Templates/DownloadIndicatorTemplate.xaml" />
+            </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>
+    </UserControl.Resources>
+    <Grid Margin="6">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition />
+        </Grid.ColumnDefinitions>
+        <TextBlock Grid.Row="0" Text="Current Version:" TextAlignment="Right" FontWeight="SemiBold" Margin="2" />
+        <TextBlock Grid.Row="0" Grid.Column="1" Margin="2" Text="{Binding Source={x:Static editor:App.Version}, StringFormat='v{0}'}"/>
+        <TextBlock Grid.Row="1" Text="Target Version:" TextAlignment="Right" FontWeight="SemiBold" Margin="2" />
+        <TextBlock Grid.Row="1" Grid.Column="1" Margin="2" Text="{Binding Target.Version, StringFormat='v{0}'}" />
+        <TextBlock Grid.Row="2" Text="Release Notes:" TextAlignment="Right" FontWeight="SemiBold" Margin="2" VerticalAlignment="Top" />
+        <RichTextBox Grid.Row="2" Grid.Column="1" FocusVisualStyle="{x:Null}" IsReadOnly="True" Margin="2" VerticalScrollBarVisibility="Auto" Loaded="RichTextBox_Loaded" />
+        <StackPanel Grid.Row="3" Grid.ColumnSpan="2">
+            <StackPanel.Resources>
+                <local:ByteGroupsConverter x:Key="ByteGroupsConverter" />
+                <constructs:EqualityConverter x:Key="EqualityConverter" />
+                <Style TargetType="{x:Type Button}" BasedOn="{StaticResource {x:Type Button}}">
+                    <Setter Property="FocusVisualStyle" Value="{x:Null}" />
+                    <Setter Property="MinWidth" Value="175" />
+                    <Setter Property="Margin" Value="2" />
+                </Style>
+            </StackPanel.Resources>
+            <StackPanel Margin="2">
+                <StackPanel.Style>
+                    <Style TargetType="{x:Type StackPanel}">
+                        <Style.Triggers>
+                            <DataTrigger Binding="{Binding BytesTransferred}" Value="{x:Null}">
+                                <Setter Property="Visibility" Value="Collapsed" />
+                            </DataTrigger>
+                        </Style.Triggers>
+                    </Style>
+                </StackPanel.Style>
+                <constructs:StatefulProgressBar Minimum="0" Maximum="{Binding Target.ZipFileSize}" Height="24" Value="{Binding BytesTransferred, Mode=OneWay, TargetNullValue=0}">
+                    <constructs:StatefulProgressBar.Style>
+                        <Style TargetType="{x:Type constructs:StatefulProgressBar}">
+                            <Style.Triggers>
+                                <Trigger Property="Value" Value="0">
+                                    <Setter Property="IsIndeterminate" Value="True" />
+                                </Trigger>
+                                <DataTrigger Binding="{Binding Error}" Value="{x:Null}">
+                                    <Setter Property="State" Value="Default" />
+                                </DataTrigger>
+                            </Style.Triggers>
+                            <Setter Property="State" Value="Error" />
+                        </Style>
+                    </constructs:StatefulProgressBar.Style>
+                </constructs:StatefulProgressBar>
+                <TextBlock Margin="0,2">
+                    <TextBlock.Style>
+                        <Style TargetType="{x:Type TextBlock}">
+                            <Style.Triggers>
+                                <DataTrigger Binding="{Binding Error, Converter={StaticResource EqualityConverter}}" Value="False">
+                                    <Setter Property="Text" Value="{Binding Error.Message, StringFormat='Error: {0}'}" />
+                                    <Setter Property="Foreground" Value="Red" />
+                                </DataTrigger>
+                                <DataTrigger Value="True">
+                                    <DataTrigger.Binding>
+                                        <MultiBinding Converter="{StaticResource EqualityConverter}">
+                                            <Binding Path="BytesTransferred" />
+                                            <Binding Path="Target.ZipFileSize" />
+                                        </MultiBinding>
+                                    </DataTrigger.Binding>
+                                    <Setter Property="Text" Value="Preparing to update..." />
+                                </DataTrigger>
+                            </Style.Triggers>
+                            <Setter Property="Text">
+                                <Setter.Value>
+                                    <MultiBinding StringFormat="Transferred: {0} out of {1}; Current Speed: {2}/s">
+                                        <Binding Path="BytesTransferred" Converter="{StaticResource ByteGroupsConverter}" />
+                                        <Binding Path="Target.ZipFileSize" Converter="{StaticResource ByteGroupsConverter}" />
+                                        <Binding Path="Speed" Converter="{StaticResource ByteGroupsConverter}" />
+                                    </MultiBinding>
+                                </Setter.Value>
+                            </Setter>
+                        </Style>
+                    </TextBlock.Style>
+                </TextBlock>
+            </StackPanel>
+            <StackPanel Margin="2" HorizontalAlignment="Center" Orientation="Horizontal">
+                <Button Command="{Binding DownloadCommand}" IsDefault="True">
+                    <StackPanel Orientation="Horizontal">
+                        <ContentPresenter Content="-" ContentTemplate="{StaticResource DownloadIndicatorTemplate}" VerticalAlignment="Center">
+                            <ContentPresenter.Style>
+                                <Style TargetType="{x:Type ContentPresenter}">
+                                    <Style.Triggers>
+                                        <DataTrigger Binding="{Binding IsEnabled, RelativeSource={RelativeSource AncestorType={x:Type Button}}}" Value="False">
+                                            <Setter Property="Opacity" Value="0.25" />
+                                        </DataTrigger>
+                                    </Style.Triggers>
+                                </Style>
+                            </ContentPresenter.Style>
+                        </ContentPresenter>
+                        <TextBlock Text="Download" VerticalAlignment="Center" Margin="2,0" />
+                    </StackPanel>
+                </Button>
+                <Button Content="Close" Command="{Binding CancelCommand}" />
+            </StackPanel>
+        </StackPanel>
+    </Grid>
+</UserControl>
